requirements.txt
Django>=4.2,<5
djangorestframework
graphene-django
psycopg2-binary
pytest
pytest-django
pytest-cov
factory-boy
djangorestframework-simplejwt
coverage
dj-database-url
django-redis
django-cors-headers
gunicorn
whitenoise
celery
redis
ruff
requests
python-dotenv
